+++ main4366.cpp
@@ -152,7 +152,7 @@ case 1:
 ncD4D = ab3jMG2(-187068881, 1389346315, (short)-629127697, (unsigned short)1307555391);
 break;
 case 2:
-Hj9DA = hiwxQ9L3c(438111771, 1396719498, vks0zEvaxD, ms08L, sy6__C);
+Hj9DA = hiwxQ9L3c((unsigned short)438111771, 1396719498, vks0zEvaxD, ms08L, sy6__C);
 break;
 case 3:
 wbmD34o /= 412130488;
